Refactor Sequence Paxos into an own module
Refactor index variables to have more intuitive names
