+++ omnipaxos_core/src/lib.rs
@@ -17,8 +17,9 @@ pub mod ballot_leader_election;
 pub mod messages;
 /// The user-facing Omni-Paxos struct.
 pub mod omni_paxos;
+pub(crate) mod sequence_paxos;
 /// The core replication algorithm of Omni-Paxos.
-pub mod sequence_paxos;
+// pub mod sequence_paxos;
 /// Traits and structs related to the backend storage of an Omni-Paxos replica.
 pub mod storage;
 /// A module containing helper functions and structs.
