+++ omnipaxos_core/src/omni_paxos.rs
@@ -1,13 +1,125 @@
+#[cfg(feature = "hocon_config")]
+use crate::utils::hocon_kv::*;
 use crate::{
-    ballot_leader_election::{BLEConfig, Ballot, BallotLeaderElection},
+    ballot_leader_election::{Ballot, BallotLeaderElection},
     messages::Message,
-    sequence_paxos::{
-        CompactionErr, ProposeErr, ReconfigurationRequest, SequencePaxos, SequencePaxosConfig,
-    },
+    sequence_paxos::{CompactionErr, ProposeErr, SequencePaxos},
     storage::{Entry, Snapshot, StopSign, Storage},
-    util::{LogEntry, NodeId},
+    util::{
+        defaults::{BUFFER_SIZE, HB_TIMEOUT},
+        LogEntry, NodeId,
+    },
 };
+#[cfg(feature = "hocon_config")]
+use hocon::Hocon;
+#[cfg(feature = "logging")]
+use slog::Logger;
 use std::ops::RangeBounds;
+
+/// Configuration for `OmniPaxos`.
+/// # Fields
+/// * `configuration_id`: The identifier for the configuration that this Sequence Paxos replica is part of.
+/// * `pid`: The unique identifier of this node. Must not be 0.
+/// * `peers`: The peers of this node i.e. the `pid`s of the other replicas in the configuration.
+/// * `buffer_size`: The buffer size for outgoing messages.
+/// * `skip_prepare_use_leader`: The initial leader of the cluster. Could be used in combination with reconfiguration to skip the prepare phase in the new configuration.
+/// * `logger`: Custom logger for logging events of Sequence Paxos.
+/// * `logger_file_path`: The path where the default logger logs events.
+#[allow(missing_docs)]
+#[derive(Clone, Debug)]
+pub struct OmniPaxosConfig {
+    pub configuration_id: u32,
+    pub pid: NodeId,
+    pub peers: Vec<u64>,
+    pub buffer_size: usize,
+    pub skip_prepare_use_leader: Option<Ballot>,
+    pub logger_file_path: Option<String>,
+    /*** BLE config fields ***/
+    pub leader_priority: u64,
+    pub hb_delay: u64,
+    pub initial_leader: Option<Ballot>,
+    pub initial_delay: Option<u64>,
+    #[cfg(feature = "logging")]
+    pub logger_path: Option<String>,
+}
+
+impl OmniPaxosConfig {
+    /// Creates a new `OmniPaxosConfig` from a `Hocon` object.
+    #[cfg(feature = "hocon_config")]
+    pub fn with_hocon(h: &Hocon) -> Self {
+        let mut config = Self::default();
+        config.configuration_id = h[CONFIG_ID].as_i64().expect("Failed to load config ID") as u32;
+        config.pid = h[PID].as_i64().expect("Failed to load PID") as u64;
+        match &h[PEERS] {
+            Hocon::Array(v) => {
+                let peers = v
+                    .iter()
+                    .map(|x| x.as_i64().expect("Failed to load pid in Hocon array") as u64)
+                    .collect();
+                config.peers = peers;
+            }
+            _ => {
+                unimplemented!("Peers in Hocon should be parsed as array!")
+            }
+        }
+        if let Some(b) = h[BUFFER_SIZE].as_i64() {
+            config.buffer_size = b as usize;
+        }
+        if let Some(p) = h[PRIORITY].as_i64().map(|p| p as u64) {
+            config.leader_priority = p;
+        }
+        config.initial_delay = h[INITIAL_DELAY].as_i64().map(|i| i as u64);
+        config.hb_delay = h[HB_DELAY]
+            .as_i64()
+            .expect("Failed to load heartbeat delay") as u64;
+
+        config.logger_file_path = h[LOG_FILE_PATH].as_string();
+        config
+    }
+
+    /// Checks all configurations and returns the local OmniPaxos node if successful.
+    pub fn build<T, S, B>(self, storage: B) -> OmniPaxos<T, S, B>
+    where
+        T: Entry,
+        S: Snapshot<T>,
+        B: Storage<T, S>,
+    {
+        assert_ne!(self.pid, 0, "Pid cannot be 0");
+        assert_ne!(self.configuration_id, 0, "Configuration id cannot be 0");
+        assert!(!self.peers.is_empty(), "Peers cannot be empty");
+        assert!(
+            !self.peers.contains(&self.pid),
+            "Peers should not include self pid"
+        );
+        assert!(self.buffer_size > 0, "Buffer size must be greater than 0");
+        if let Some(x) = self.skip_prepare_use_leader {
+            assert_ne!(x.pid, 0, "Initial leader cannot be 0")
+        };
+        OmniPaxos {
+            seq_paxos: SequencePaxos::with(self.clone().into(), storage),
+            ble: BallotLeaderElection::with(self.into()),
+        }
+    }
+}
+
+impl Default for OmniPaxosConfig {
+    fn default() -> Self {
+        Self {
+            configuration_id: 0,
+            pid: 0,
+            peers: Vec::new(),
+            buffer_size: BUFFER_SIZE,
+            skip_prepare_use_leader: None,
+            logger_file_path: None,
+            leader_priority: 0,
+            hb_delay: HB_TIMEOUT,
+            initial_leader: None,
+            initial_delay: None,
+            #[cfg(feature = "logging")]
+            logger_path: None,
+        }
+    }
+}
 
 /// The `OmniPaxos` struct represents an OmniPaxos server. Maintains the replicated log that can be read from and appended to.
 /// It also handles incoming messages and produces outgoing messages that you need to fetch and send periodically using your own network implementation.
@@ -27,14 +139,6 @@ where
     S: Snapshot<T>,
     B: Storage<T, S>,
 {
-    /// Creates an `OmniPaxos` server
-    pub fn with(sp_config: SequencePaxosConfig, ble_config: BLEConfig, storage: B) -> Self {
-        Self {
-            seq_paxos: SequencePaxos::with(sp_config, storage),
-            ble: BallotLeaderElection::with(ble_config),
-        }
-    }
-
     /// Initiates the trim process.
     /// # Arguments
     /// * `trim_index` - Deletes all entries up to [`trim_index`], if the [`trim_index`] is `None` then the minimum index accepted by **ALL** servers will be used as the [`trim_index`].
@@ -163,6 +267,28 @@ where
     pub fn tick(&mut self) {
         if let Some(b) = self.ble.tick() {
             self.seq_paxos.handle_leader(b);
+        }
+    }
+}
+
+/// Used for proposing reconfiguration of the cluster.
+#[derive(Debug, Clone)]
+pub struct ReconfigurationRequest {
+    /// The id of the servers in the new configuration.
+    pub(crate) new_configuration: Vec<u64>,
+    /// Optional metadata to be decided with the reconfiguration.
+    pub(crate) metadata: Option<Vec<u8>>,
+}
+
+impl ReconfigurationRequest {
+    /// create a `ReconfigurationRequest`.
+    /// # Arguments
+    /// * new_configuration: The pids of the nodes in the new configuration.
+    /// * metadata: Some optional metadata in raw bytes. This could include some auxiliary data for the new configuration to start with.
+    pub fn with(new_configuration: Vec<u64>, metadata: Option<Vec<u8>>) -> Self {
+        Self {
+            new_configuration,
+            metadata,
         }
     }
 }
