+++ omnipaxos_core/src/util.rs
@@ -9,16 +9,20 @@ use std::{cmp::Ordering, fmt::Debug, marker::PhantomData};
 /// Promise without the suffix
 pub(crate) struct PromiseMetaData {
     pub n: Ballot,
-    pub la: u64,
+    pub accepted_idx: u64,
     pub pid: NodeId,
     pub stopsign: Option<StopSign>,
 }
 
 impl PartialOrd for PromiseMetaData {
     fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        let ordering = if self.n == other.n && self.la == other.la && self.pid == other.pid {
+        let ordering = if self.n == other.n
+            && self.accepted_idx == other.accepted_idx
+            && self.pid == other.pid
+        {
             Ordering::Equal
-        } else if self.n > other.n || (self.n == other.n && self.la > other.la) {
+        } else if self.n > other.n || (self.n == other.n && self.accepted_idx > other.accepted_idx)
+        {
             Ordering::Greater
         } else {
             Ordering::Less
@@ -29,7 +33,7 @@ impl PartialOrd for PromiseMetaData {
 
 impl PartialEq for PromiseMetaData {
     fn eq(&self, other: &Self) -> bool {
-        self.n == other.n && self.la == other.la && self.pid == other.pid
+        self.n == other.n && self.accepted_idx == other.accepted_idx && self.pid == other.pid
     }
 }
 
@@ -41,9 +45,9 @@ where
 {
     pub n_leader: Ballot,
     pub promises_meta: Vec<Option<PromiseMetaData>>,
-    pub las: Vec<u64>,
-    pub lds: Vec<Option<u64>>,
-    pub lc: u64, // length of longest chosen seq
+    pub accepted_indexes: Vec<u64>,
+    pub decided_indexes: Vec<Option<u64>>,
+    pub chosen_idx: u64, // length of longest chosen seq
     pub max_promise_meta: PromiseMetaData,
     pub max_promise: Option<(Option<SnapshotType<T, S>>, Vec<T>)>, // (decided_snapshot, suffix)
     #[cfg(feature = "batch_accept")]
@@ -60,16 +64,16 @@ where
 {
     pub fn with(
         n_leader: Ballot,
-        lds: Option<Vec<Option<u64>>>,
+        decided_indexes: Option<Vec<Option<u64>>>,
         max_pid: usize,
         majority: usize,
     ) -> Self {
         Self {
             n_leader,
             promises_meta: vec![None; max_pid],
-            las: vec![0; max_pid],
-            lds: lds.unwrap_or_else(|| vec![None; max_pid]),
-            lc: 0,
+            accepted_indexes: vec![0; max_pid],
+            decided_indexes: decided_indexes.unwrap_or_else(|| vec![None; max_pid]),
+            chosen_idx: 0,
             max_promise_meta: PromiseMetaData::default(),
             max_promise: None,
             #[cfg(feature = "batch_accept")]
@@ -85,13 +89,13 @@ where
     }
 
     pub fn set_decided_idx(&mut self, pid: NodeId, idx: Option<u64>) {
-        self.lds[Self::pid_to_idx(pid)] = idx;
+        self.decided_indexes[Self::pid_to_idx(pid)] = idx;
     }
 
     pub fn set_promise(&mut self, prom: Promise<T, S>, from: u64, check_max_prom: bool) -> bool {
         let promise_meta = PromiseMetaData {
             n: prom.n_accepted,
-            la: prom.la,
+            accepted_idx: prom.accepted_idx,
             pid: from,
             stopsign: prom.stopsign,
         };
@@ -99,7 +103,7 @@ where
             self.max_promise_meta = promise_meta.clone();
             self.max_promise = Some((prom.decided_snapshot, prom.suffix))
         }
-        self.lds[Self::pid_to_idx(from)] = Some(prom.ld);
+        self.decided_indexes[Self::pid_to_idx(from)] = Some(prom.decided_idx);
         self.promises_meta[Self::pid_to_idx(from)] = Some(promise_meta);
         let num_promised = self.promises_meta.iter().filter(|x| x.is_some()).count();
         num_promised >= self.majority
@@ -124,7 +128,7 @@ where
     }
 
     pub fn get_min_all_accepted_idx(&self) -> &u64 {
-        self.las
+        self.accepted_indexes
             .iter()
             .min()
             .expect("Should be all initialised to 0!")
@@ -136,15 +140,15 @@ where
     }
 
     pub fn set_chosen_idx(&mut self, idx: u64) {
-        self.lc = idx;
+        self.chosen_idx = idx;
     }
 
     pub fn get_chosen_idx(&self) -> u64 {
-        self.lc
+        self.chosen_idx
     }
 
     pub fn get_promised_followers(&self) -> Vec<u64> {
-        self.lds
+        self.decided_indexes
             .iter()
             .enumerate()
             .filter(|(pid, x)| x.is_some() && *pid != Self::pid_to_idx(self.n_leader.pid))
@@ -159,7 +163,7 @@ where
     }
 
     pub fn set_accepted_idx(&mut self, pid: NodeId, idx: u64) {
-        self.las[Self::pid_to_idx(pid)] = idx;
+        self.accepted_indexes[Self::pid_to_idx(pid)] = idx;
     }
 
     #[cfg(feature = "batch_accept")]
@@ -172,7 +176,7 @@ where
     }
 
     pub fn get_decided_idx(&self, pid: NodeId) -> &Option<u64> {
-        self.lds.get(Self::pid_to_idx(pid)).unwrap()
+        self.decided_indexes.get(Self::pid_to_idx(pid)).unwrap()
     }
 
     pub fn is_stopsign_chosen(&self) -> bool {
@@ -181,7 +185,11 @@ where
     }
 
     pub fn is_chosen(&self, idx: u64) -> bool {
-        self.las.iter().filter(|la| **la >= idx).count() >= self.majority
+        self.accepted_indexes
+            .iter()
+            .filter(|la| **la >= idx)
+            .count()
+            >= self.majority
     }
 
     pub fn take_max_promise_stopsign(&mut self) -> Option<StopSign> {
